Выносит логику из main в trip-controller
