+++ markup/list-sort.html
@@ -213,7 +213,8 @@
                       <span class="visually-hidden">Open event</span>
                     </button>
                   </div>
-                </li><li class="trip-events__item">
+                </li>
+                <li class="trip-events__item">
                   <div class="event">
                     <div class="event__type">
                       <img class="event__type-icon" width="42" height="42" src="img/icons/sightseeing.png" alt="Event type icon">
